requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
Brotli==1.1.0
cffi==1.16.0
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.1.7
django-filter==23.5
django-import-export==3.3.7
django-redisboard==8.4.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
fonttools==4.47.2
html5lib==1.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
pillow==10.2.0
pycparser==2.21
pydyf==0.8.0
pyphen==0.14.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
tinycss2==1.2.1
typing_extensions==4.9.0
weasyprint==60.2
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
zopfli==0.2.3
